[MS-68] Refactor: 리뷰 반영
- 중복 의존성 주입 수정
- 에러 메세지 수정
- sessionId
- 의미없는 형변환 제거
- toEntity함수를 Mapper 클래스 내부로 이동
