+++ src/main/java/com/modutaxi/api/common/exception/errorcode/ChatErrorCode.java
@@ -10,9 +10,9 @@ import org.springframework.http.HttpStatus;
 @RequiredArgsConstructor
 public enum ChatErrorCode implements ErrorCode {
     FULL_CHAT_ROOM("CHAT_001", "방이 꽉 차있습니다", HttpStatus.BAD_REQUEST),
-    FAULT_ROOM_ID("CHAT_002", "잘못된 방", HttpStatus.BAD_REQUEST),
-    ALREADY_ROOM_IN("CHAT_003", "이미 방에 참가해있잖아~", HttpStatus.BAD_REQUEST),
-    ALREADY_ROOM_OUT("CHAT_004", "이미 나가있잖아 인마", HttpStatus.BAD_REQUEST),
+    FAULT_ROOM_ID("CHAT_002", "잘못된 방입니다. 다시 확인하세요.", HttpStatus.BAD_REQUEST),
+    ALREADY_ROOM_IN("CHAT_003", "당신은 이미 방에 참가해있습니다.", HttpStatus.BAD_REQUEST),
+    ALREADY_ROOM_OUT("CHAT_004", "방에서 이미 나간 상태입니다.", HttpStatus.BAD_REQUEST),
     ;
 
     private final String errorCode;
