+++ src/main/java/com/modutaxi/api/domain/chatmessage/entity/ChatMessage.java
@@ -39,15 +39,4 @@ public class ChatMessage extends BaseTime {
     private String memberId;
 
     private LocalDateTime dateTime;
-
-    public static ChatMessage toEntity(ChatMessageRequestDto messageRequestDto, Room room){
-        return ChatMessage.builder()
-                .room(room)
-                .messageType(messageRequestDto.getType())
-                .content(messageRequestDto.getContent())
-                .sender(messageRequestDto.getSender())
-                .memberId(messageRequestDto.getMemberId())
-                .dateTime(messageRequestDto.getDateTime())
-                .build();
-    }
 }
